# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "neuroagent-framework",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tools/{tool_id}": {
      "get": {
        "tags": [
          "tools"
        ],
        "summary": "Get Tool",
        "operationId": "get_tool_tools__tool_id__get",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Tool Tools  Tool Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tools/{tool_id}/test": {
      "post": {
        "tags": [
          "tools"
        ],
        "summary": "Test Tool",
        "operationId": "test_tool_tools__tool_id__test_post",
        "parameters": [
          {
            "name": "tool_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tool Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolTestRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Test Tool Tools  Tool Id  Test Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/use-cases/{use_case_id}": {
      "get": {
        "tags": [
          "use-cases"
        ],
        "summary": "Get Use Case",
        "operationId": "get_use_case_use_cases__use_case_id__get",
        "parameters": [
          {
            "name": "use_case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Use Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Use Case Use Cases  Use Case Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/use-cases/{use_case_id}/run": {
      "post": {
        "tags": [
          "use-cases"
        ],
        "summary": "Run Use Case",
        "operationId": "run_use_case_use_cases__use_case_id__run_post",
        "parameters": [
          {
            "name": "use_case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Use Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Run Use Case Use Cases  Use Case Id  Run Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ToolTestRequest": {
        "properties": {
          "input_payload": {
            "type": "object",
            "title": "Input Payload"
          }
        },
        "type": "object",
        "title": "ToolTestRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}